refactor: 메뉴(Menu)와 메뉴 그룹(MenuGroup)를 기반으로 패키지 분리
