+++ src/main/java/kitchenpos/deliveryorders/infra/KitchenridersClient.java
@@ -1,0 +1,8 @@
+package kitchenpos.deliveryorders.infra;
+
+import java.math.BigDecimal;
+import java.util.UUID;
+
+public interface KitchenridersClient {
+    void requestDelivery(UUID orderId, BigDecimal amount, String deliveryAddress);
+}
