--- src/main/java/kitchenpos/domain/JpaMenuGroupRepository.java
@@ -1,8 +1,0 @@
-package kitchenpos.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface JpaMenuGroupRepository extends MenuGroupRepository, JpaRepository<MenuGroup, UUID> {
-}
