+++ src/main/java/kitchenpos/eatinorders/application/OrderService.java
@@ -1,0 +1,193 @@
+package kitchenpos.eatinorders.application;
+
+import kitchenpos.deliveryorders.infra.KitchenridersClient;
+import kitchenpos.eatinorders.domain.Order;
+import kitchenpos.eatinorders.domain.OrderLineItem;
+import kitchenpos.eatinorders.domain.OrderRepository;
+import kitchenpos.eatinorders.domain.OrderStatus;
+import kitchenpos.eatinorders.domain.OrderTable;
+import kitchenpos.eatinorders.domain.OrderTableRepository;
+import kitchenpos.eatinorders.domain.OrderType;
+import kitchenpos.menus.domain.Menu;
+import kitchenpos.menus.domain.MenuRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.Objects;
+import java.util.UUID;
+
+@Service
+public class OrderService {
+    private final OrderRepository orderRepository;
+    private final MenuRepository menuRepository;
+    private final OrderTableRepository orderTableRepository;
+    private final KitchenridersClient kitchenridersClient;
+
+    public OrderService(
+            final OrderRepository orderRepository,
+            final MenuRepository menuRepository,
+            final OrderTableRepository orderTableRepository,
+            final KitchenridersClient kitchenridersClient
+    ) {
+        this.orderRepository = orderRepository;
+        this.menuRepository = menuRepository;
+        this.orderTableRepository = orderTableRepository;
+        this.kitchenridersClient = kitchenridersClient;
+    }
+
+    @Transactional
+    public Order create(final Order request) {
+        final OrderType type = request.getType();
+        if (Objects.isNull(type)) {
+            throw new IllegalArgumentException("주문 유형이 존재해야 합니다");
+        }
+        final List<OrderLineItem> orderLineItemRequests = request.getOrderLineItems();
+        if (Objects.isNull(orderLineItemRequests) || orderLineItemRequests.isEmpty()) {
+            throw new IllegalArgumentException("주문 항목이 존재해야 합니다.");
+        }
+        final List<Menu> menus = menuRepository.findAllByIdIn(
+                orderLineItemRequests.stream()
+                        .map(OrderLineItem::getMenuId)
+                        .toList()
+        );
+        if (menus.size() != orderLineItemRequests.size()) {
+            throw new IllegalArgumentException();
+        }
+        final List<OrderLineItem> orderLineItems = new ArrayList<>();
+        for (final OrderLineItem orderLineItemRequest : orderLineItemRequests) {
+            final long quantity = orderLineItemRequest.getQuantity();
+            if (type != OrderType.EAT_IN) {
+                if (quantity < 0) {
+                    throw new IllegalArgumentException("주문 유형이 `배달` 혹은 `테이크아웃`시 주문 수량이 0보다 커야 합니다");
+                }
+            }
+            final Menu menu = menuRepository.findById(orderLineItemRequest.getMenuId())
+                    .orElseThrow(NoSuchElementException::new);
+            if (!menu.isDisplayed()) {
+                throw new IllegalStateException("메뉴가 표시되어야 주문이 가능합니다.");
+            }
+            if (menu.getPrice().compareTo(orderLineItemRequest.getPrice()) != 0) {
+                throw new IllegalArgumentException("주문항목에 있는 가격이 메뉴에 있는 가격과 동일해야 합니다.");
+            }
+            final OrderLineItem orderLineItem = new OrderLineItem();
+            orderLineItem.setMenu(menu);
+            orderLineItem.setQuantity(quantity);
+            orderLineItems.add(orderLineItem);
+        }
+        Order order = new Order();
+        order.setId(UUID.randomUUID());
+        order.setType(type);
+        order.setStatus(OrderStatus.WAITING);
+        order.setOrderDateTime(LocalDateTime.now());
+        order.setOrderLineItems(orderLineItems);
+        if (type == OrderType.DELIVERY) {
+            final String deliveryAddress = request.getDeliveryAddress();
+            if (Objects.isNull(deliveryAddress) || deliveryAddress.isEmpty()) {
+                throw new IllegalArgumentException("주문 유형이 `배달`이면 배달에 필요한 주소가 존재해야 한다.");
+            }
+            order.setDeliveryAddress(deliveryAddress);
+        }
+        if (type == OrderType.EAT_IN) {
+            final OrderTable orderTable = orderTableRepository.findById(request.getOrderTableId())
+                    .orElseThrow(NoSuchElementException::new);
+            if (!orderTable.isOccupied()) {
+                throw new IllegalStateException("주문 유형이 `매장 내 식사`이면 가게 안에 자리가 꽉 차지 않아야 한다.");
+            }
+            order.setOrderTable(orderTable);
+        }
+        return orderRepository.save(order);
+    }
+
+    @Transactional
+    public Order accept(final UUID orderId) {
+        final Order order = orderRepository.findById(orderId)
+                .orElseThrow(NoSuchElementException::new);
+        if (order.getStatus() != OrderStatus.WAITING) {
+            throw new IllegalStateException();
+        }
+        if (order.getType() == OrderType.DELIVERY) {
+            BigDecimal sum = BigDecimal.ZERO;
+            for (final OrderLineItem orderLineItem : order.getOrderLineItems()) {
+                sum = orderLineItem.getMenu()
+                        .getPrice()
+                        .multiply(BigDecimal.valueOf(orderLineItem.getQuantity()));
+            }
+            kitchenridersClient.requestDelivery(orderId, sum, order.getDeliveryAddress());
+        }
+        order.setStatus(OrderStatus.ACCEPTED);
+        return order;
+    }
+
+    @Transactional
+    public Order serve(final UUID orderId) {
+        final Order order = orderRepository.findById(orderId)
+                .orElseThrow(NoSuchElementException::new);
+        if (order.getStatus() != OrderStatus.ACCEPTED) {
+            throw new IllegalStateException();
+        }
+        order.setStatus(OrderStatus.SERVED);
+        return order;
+    }
+
+    @Transactional
+    public Order startDelivery(final UUID orderId) {
+        final Order order = orderRepository.findById(orderId)
+                .orElseThrow(NoSuchElementException::new);
+        if (order.getType() != OrderType.DELIVERY) {
+            throw new IllegalStateException();
+        }
+        if (order.getStatus() != OrderStatus.SERVED) {
+            throw new IllegalStateException();
+        }
+        order.setStatus(OrderStatus.DELIVERING);
+        return order;
+    }
+
+    @Transactional
+    public Order completeDelivery(final UUID orderId) {
+        final Order order = orderRepository.findById(orderId)
+                .orElseThrow(NoSuchElementException::new);
+        if (order.getStatus() != OrderStatus.DELIVERING) {
+            throw new IllegalStateException();
+        }
+        order.setStatus(OrderStatus.DELIVERED);
+        return order;
+    }
+
+    @Transactional
+    public Order complete(final UUID orderId) {
+        final Order order = orderRepository.findById(orderId)
+                .orElseThrow(NoSuchElementException::new);
+        final OrderType type = order.getType();
+        final OrderStatus status = order.getStatus();
+        if (type == OrderType.DELIVERY) {
+            if (status != OrderStatus.DELIVERED) {
+                throw new IllegalStateException();
+            }
+        }
+        if (type == OrderType.TAKEOUT || type == OrderType.EAT_IN) {
+            if (status != OrderStatus.SERVED) {
+                throw new IllegalStateException();
+            }
+        }
+        order.setStatus(OrderStatus.COMPLETED);
+        if (type == OrderType.EAT_IN) {
+            final OrderTable orderTable = order.getOrderTable();
+            if (!orderRepository.existsByOrderTableAndStatusNot(orderTable, OrderStatus.COMPLETED)) {
+                orderTable.setNumberOfGuests(0);
+                orderTable.setOccupied(false);
+            }
+        }
+        return order;
+    }
+
+    @Transactional(readOnly = true)
+    public List<Order> findAll() {
+        return orderRepository.findAll();
+    }
+}
