+++ src/main/java/kitchenpos/eatinorders/domain/JpaOrderRepository.java
@@ -1,0 +1,8 @@
+package kitchenpos.eatinorders.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface JpaOrderRepository extends OrderRepository, JpaRepository<Order, UUID> {
+}
