+++ src/main/java/kitchenpos/menus/application/MenuGroupService.java
@@ -1,0 +1,36 @@
+package kitchenpos.menus.application;
+
+import kitchenpos.menus.domain.MenuGroup;
+import kitchenpos.menus.domain.MenuGroupRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.Objects;
+import java.util.UUID;
+
+@Service
+public class MenuGroupService {
+    private final MenuGroupRepository menuGroupRepository;
+
+    public MenuGroupService(final MenuGroupRepository menuGroupRepository) {
+        this.menuGroupRepository = menuGroupRepository;
+    }
+
+    @Transactional
+    public MenuGroup create(final MenuGroup request) {
+        final String name = request.getName();
+        if (Objects.isNull(name) || name.isEmpty()) {
+            throw new IllegalArgumentException("메뉴 그룹의 이름이 존재해야 합니다.");
+        }
+        final MenuGroup menuGroup = new MenuGroup();
+        menuGroup.setId(UUID.randomUUID());
+        menuGroup.setName(name);
+        return menuGroupRepository.save(menuGroup);
+    }
+
+    @Transactional(readOnly = true)
+    public List<MenuGroup> findAll() {
+        return menuGroupRepository.findAll();
+    }
+}
