+++ src/main/java/kitchenpos/menus/application/MenuService.java
@@ -1,0 +1,155 @@
+package kitchenpos.menus.application;
+
+import io.micrometer.common.util.StringUtils;
+import kitchenpos.menus.domain.Menu;
+import kitchenpos.menus.domain.MenuGroup;
+import kitchenpos.menus.domain.MenuGroupRepository;
+import kitchenpos.menus.domain.MenuProduct;
+import kitchenpos.menus.domain.MenuRepository;
+import kitchenpos.products.domain.Product;
+import kitchenpos.products.domain.ProductRepository;
+import kitchenpos.products.infra.PurgomalumClient;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.math.BigDecimal;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.Objects;
+import java.util.UUID;
+
+@Service
+public class MenuService {
+    private final MenuRepository menuRepository;
+    private final MenuGroupRepository menuGroupRepository;
+    private final ProductRepository productRepository;
+    private final PurgomalumClient purgomalumClient;
+
+    public MenuService(
+            final MenuRepository menuRepository,
+            final MenuGroupRepository menuGroupRepository,
+            final ProductRepository productRepository,
+            final PurgomalumClient purgomalumClient
+    ) {
+        this.menuRepository = menuRepository;
+        this.menuGroupRepository = menuGroupRepository;
+        this.productRepository = productRepository;
+        this.purgomalumClient = purgomalumClient;
+    }
+
+    @Transactional
+    public Menu create(final Menu request) {
+        final BigDecimal price = request.getPrice();
+        if (Objects.isNull(price) || price.compareTo(BigDecimal.ZERO) < 0) {
+            throw new IllegalArgumentException("가격이 존재하거나 0원 이상이어야 합니다.");
+        }
+        final MenuGroup menuGroup = menuGroupRepository.findById(request.getMenuGroupId())
+                .orElseThrow(NoSuchElementException::new);
+        final List<MenuProduct> menuProductRequests = request.getMenuProducts();
+        if (Objects.isNull(menuProductRequests) || menuProductRequests.isEmpty()) {
+            throw new IllegalArgumentException("메뉴에 포함된 상품이 존재해야 합니다.");
+        }
+        final List<Product> products = productRepository.findAllByIdIn(
+                menuProductRequests.stream()
+                        .map(MenuProduct::getProductId)
+                        .toList()
+        );
+        if (products.size() != menuProductRequests.size()) {
+            throw new IllegalArgumentException();
+        }
+        final List<MenuProduct> menuProducts = new ArrayList<>();
+        BigDecimal sum = BigDecimal.ZERO;
+        for (final MenuProduct menuProductRequest : menuProductRequests) {
+            final long quantity = menuProductRequest.getQuantity();
+            if (quantity < 0) {
+                throw new IllegalArgumentException();
+            }
+            final Product product = productRepository.findById(menuProductRequest.getProductId())
+                    .orElseThrow(NoSuchElementException::new);
+            sum = sum.add(
+                    product.getPrice()
+                            .multiply(BigDecimal.valueOf(quantity))
+            );
+            final MenuProduct menuProduct = new MenuProduct();
+            menuProduct.setProduct(product);
+            menuProduct.setQuantity(quantity);
+            menuProducts.add(menuProduct);
+        }
+        if (price.compareTo(sum) > 0) {
+            throw new IllegalArgumentException();
+        }
+        final String name = request.getName();
+        if (StringUtils.isBlank(name) || purgomalumClient.containsProfanity(name)) {
+            throw new IllegalArgumentException("올바른 메뉴 이름을 입력해야 합니다.");
+        }
+        final Menu menu = new Menu();
+        menu.setId(UUID.randomUUID());
+        menu.setName(name);
+        menu.setPrice(price);
+        menu.setMenuGroup(menuGroup);
+        menu.setDisplayed(request.isDisplayed());
+        menu.setMenuProducts(menuProducts);
+        return menuRepository.save(menu);
+    }
+
+    @Transactional
+    public Menu changePrice(final UUID menuId, final Menu request) {
+        final BigDecimal price = request.getPrice();
+        if (Objects.isNull(price) || price.compareTo(BigDecimal.ZERO) < 0) {
+            throw new IllegalArgumentException("메뉴 변경시 가격이 0원 이상이어야 합니다.");
+        }
+        BigDecimal sum = BigDecimal.ZERO;
+        final Menu menu = menuRepository.findById(menuId)
+                .orElseThrow(NoSuchElementException::new);
+        for (final MenuProduct menuProduct : menu.getMenuProducts()) {
+            Product product = menuProduct.getProduct();
+            if (Objects.isNull(product) || Objects.isNull(product.getPrice())) {
+                throw new IllegalArgumentException("상품 가격이 존재해야만 합니다.");
+            }
+
+            sum = sum.add(
+                    menuProduct.getProduct()
+                            .getPrice()
+                            .multiply(BigDecimal.valueOf(menuProduct.getQuantity()))
+            );
+        }
+        if (price.compareTo(sum) > 0) {
+            throw new IllegalArgumentException("메뉴 가격은 포함된 상품 가격 합보다 클 수 없습니다.");
+        }
+        menu.setPrice(price);
+        return menu;
+    }
+
+    @Transactional
+    public Menu display(final UUID menuId) {
+        final Menu menu = menuRepository.findById(menuId)
+                .orElseThrow(() -> new NoSuchElementException("해당 ID의 메뉴가 존재하지 않습니다."));
+        BigDecimal sum = BigDecimal.ZERO;
+        for (final MenuProduct menuProduct : menu.getMenuProducts()) {
+            sum = sum.add(
+                    menuProduct.getProduct()
+                            .getPrice()
+                            .multiply(BigDecimal.valueOf(menuProduct.getQuantity()))
+            );
+        }
+        if (menu.getPrice().compareTo(sum) > 0) {
+            throw new IllegalStateException("메뉴 가격이 포함된 상품 가격보다 높아 표시할 수 없습니다.");
+        }
+        menu.setDisplayed(true);
+        return menu;
+    }
+
+    @Transactional
+    public Menu hide(final UUID menuId) {
+        final Menu menu = menuRepository.findById(menuId)
+                .orElseThrow(NoSuchElementException::new);
+        menu.setDisplayed(false);
+        return menu;
+    }
+
+    @Transactional(readOnly = true)
+    public List<Menu> findAll() {
+        return menuRepository.findAll();
+    }
+}
