+++ src/main/java/kitchenpos/menus/domain/JpaMenuGroupRepository.java
@@ -1,0 +1,8 @@
+package kitchenpos.menus.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface JpaMenuGroupRepository extends MenuGroupRepository, JpaRepository<MenuGroup, UUID> {
+}
