+++ src/main/java/kitchenpos/menus/domain/JpaMenuRepository.java
@@ -1,0 +1,14 @@
+package kitchenpos.menus.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
+
+import java.util.List;
+import java.util.UUID;
+
+public interface JpaMenuRepository extends MenuRepository, JpaRepository<Menu, UUID> {
+    @Query("select m from Menu m join m.menuProducts mp where mp.product.id = :productId")
+    @Override
+    List<Menu> findAllByProductId(@Param("productId") UUID productId);
+}
