+++ src/main/java/kitchenpos/menus/domain/MenuGroupRepository.java
@@ -1,0 +1,14 @@
+package kitchenpos.menus.domain;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+
+public interface MenuGroupRepository {
+    MenuGroup save(MenuGroup menuGroup);
+
+    Optional<MenuGroup> findById(UUID id);
+
+    List<MenuGroup> findAll();
+}
+
