+++ src/main/java/kitchenpos/menus/ui/MenuGroupRestController.java
@@ -1,0 +1,35 @@
+package kitchenpos.menus.ui;
+
+import kitchenpos.menus.application.MenuGroupService;
+import kitchenpos.menus.domain.MenuGroup;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.net.URI;
+import java.util.List;
+
+@RequestMapping("/api/menu-groups")
+@RestController
+public class MenuGroupRestController {
+    private final MenuGroupService menuGroupService;
+
+    public MenuGroupRestController(final MenuGroupService menuGroupService) {
+        this.menuGroupService = menuGroupService;
+    }
+
+    @PostMapping
+    public ResponseEntity<MenuGroup> create(@RequestBody final MenuGroup request) {
+        final MenuGroup response = menuGroupService.create(request);
+        return ResponseEntity.created(URI.create("/api/menu-groups/" + response.getId()))
+            .body(response);
+    }
+
+    @GetMapping
+    public ResponseEntity<List<MenuGroup>> findAll() {
+        return ResponseEntity.ok(menuGroupService.findAll());
+    }
+}
