+++ src/main/java/kitchenpos/menus/ui/MenuRestController.java
@@ -1,0 +1,53 @@
+package kitchenpos.menus.ui;
+
+import kitchenpos.menus.application.MenuService;
+import kitchenpos.menus.domain.Menu;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.net.URI;
+import java.util.List;
+import java.util.UUID;
+
+@RequestMapping("/api/menus")
+@RestController
+public class MenuRestController {
+    private final MenuService menuService;
+
+    public MenuRestController(final MenuService menuService) {
+        this.menuService = menuService;
+    }
+
+    @PostMapping
+    public ResponseEntity<Menu> create(@RequestBody final Menu request) {
+        final Menu response = menuService.create(request);
+        return ResponseEntity.created(URI.create("/api/menus/" + response.getId()))
+            .body(response);
+    }
+
+    @PutMapping("/{menuId}/price")
+    public ResponseEntity<Menu> changePrice(@PathVariable final UUID menuId, @RequestBody final Menu request) {
+        return ResponseEntity.ok(menuService.changePrice(menuId, request));
+    }
+
+    @PutMapping("/{menuId}/display")
+    public ResponseEntity<Menu> display(@PathVariable final UUID menuId) {
+        return ResponseEntity.ok(menuService.display(menuId));
+    }
+
+    @PutMapping("/{menuId}/hide")
+    public ResponseEntity<Menu> hide(@PathVariable final UUID menuId) {
+        return ResponseEntity.ok(menuService.hide(menuId));
+    }
+
+    @GetMapping
+    public ResponseEntity<List<Menu>> findAll() {
+        return ResponseEntity.ok(menuService.findAll());
+    }
+}
