+++ src/test/java/kitchenpos/fixtures/MenuFixtures.java
@@ -1,0 +1,93 @@
+package kitchenpos.fixtures;
+
+import static kitchenpos.fixtures.ProductFixtures.product;
+import kitchenpos.menus.domain.Menu;
+import kitchenpos.menus.domain.MenuGroup;
+import kitchenpos.menus.domain.MenuProduct;
+
+import java.math.BigDecimal;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Random;
+import java.util.UUID;
+
+public class MenuFixtures {
+
+    public static MenuProduct createMenuProduct(final UUID productId, final long quantity) {
+        MenuProduct menuProduct = new MenuProduct();
+        menuProduct.setProductId(productId);
+        menuProduct.setQuantity(quantity);
+        return menuProduct;
+    }
+
+    public static MenuGroup menuGroup() {
+        return createMenuGroup("메인메뉴");
+    }
+
+    public static MenuGroup createMenuGroup(final String name) {
+        MenuGroup menuGroup = new MenuGroup();
+        menuGroup.setId(UUID.randomUUID());
+        menuGroup.setName(name);
+        return menuGroup;
+    }
+
+
+    public static Menu createMenu(
+        final String name,
+        final BigDecimal price,
+        final boolean displayed,
+        final MenuGroup menuGroup,
+        final List<MenuProduct> menuProducts
+    ) {
+        Menu menu = new Menu();
+        menu.setId(UUID.randomUUID());
+        menu.setName(name);
+        menu.setPrice(price);
+        menu.setMenuGroup(menuGroup);
+        menu.setMenuGroupId(menuGroup.getId());
+        menu.setDisplayed(displayed);
+        menu.setMenuProducts(menuProducts);
+        return menu;
+    }
+
+    public static Menu createMenuWithoutMenuGroupId(
+        final String name,
+        final BigDecimal price,
+        final boolean displayed,
+        final List<MenuProduct> menuProducts
+    ) {
+        Menu menu  = new Menu();
+        menu.setId(UUID.randomUUID());
+        menu.setName(name);
+        menu.setPrice(price);
+        menu.setMenuGroup(null);
+        menu.setDisplayed(displayed);
+        menu.setMenuProducts(menuProducts);
+        return menu;
+    }
+
+    /**
+     * OrderServiceTest
+     */
+    public static Menu menu() {
+        return menu(19_000L, true, menuProduct());
+    }
+    public static Menu menu(final long price, final boolean displayed, final MenuProduct... menuProducts) {
+        final Menu menu = new Menu();
+        menu.setId(UUID.randomUUID());
+        menu.setName("후라이드치킨+양념치킨");
+        menu.setPrice(BigDecimal.valueOf(price));
+        menu.setMenuGroup(menuGroup());
+        menu.setDisplayed(displayed);
+        menu.setMenuProducts(Arrays.asList(menuProducts));
+        return menu;
+    }
+
+    public static MenuProduct menuProduct() {
+        final MenuProduct menuProduct = new MenuProduct();
+        menuProduct.setSeq(new Random().nextLong());
+        menuProduct.setProduct(product());
+        menuProduct.setQuantity(2L);
+        return menuProduct;
+    }
+}
