+++ src/test/java/kitchenpos/menus/application/InMemoryMenuGroupRepository.java
@@ -1,0 +1,31 @@
+package kitchenpos.menus.application;
+
+import kitchenpos.menus.domain.MenuGroup;
+import kitchenpos.menus.domain.MenuGroupRepository;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.UUID;
+
+public class InMemoryMenuGroupRepository implements MenuGroupRepository {
+    private final Map<UUID, MenuGroup> menuGroups = new HashMap<>();
+
+    @Override
+    public MenuGroup save(final MenuGroup menuGroup) {
+        menuGroups.put(menuGroup.getId(), menuGroup);
+        return menuGroup;
+    }
+
+    @Override
+    public Optional<MenuGroup> findById(final UUID id) {
+        return Optional.ofNullable(menuGroups.get(id));
+    }
+
+    @Override
+    public List<MenuGroup> findAll() {
+        return new ArrayList<>(menuGroups.values());
+    }
+}
